ADDED NEW FILE TO REPOSITORY.
time-calculator-2.0.py

This is a modified version of the previous 'time-calculator.py' file which includes error handling. This is the first version that was NOT made entirely by me, and that i had help from the AI to better understand code best practices, clean coding and better loop and functions calling management.

Also, this version is in Spanish
Signed-off-by: Hernan Moisés <hernanmartinmoises@gmail.com>