ADDED NEW DIRECTORY
time-calculator.py and time-calculator-2.0.py were moved to a new independent folder. Project is expected to grow in size and complexity.

We've also moved hours-to-HH-MM-SS.py file as well, since it was an independent python file including a function for a specific requirement that was later added to original time-calculator versions.